renaming, export IViewHandler, ViewMediator
