+++ src/robotlegs/bender/extensions/mediatorMap/api/IMediator.ts
