+++ src/robotlegs/bender/extensions/mediatorMap/api/IMediatorManager.ts
