+++ src/robotlegs/bender/extensions/mediatorMap/api/IMediatorMapping.ts
