+++ src/robotlegs/bender/extensions/mediatorMap/api/ISceneMediatorMap.ts
