+++ src/robotlegs/bender/extensions/mediatorMap/api/IViewMediatorMap.ts
