+++ src/robotlegs/bender/extensions/mediatorMap/dsl/IMediatorConfigurator.ts
