+++ src/robotlegs/bender/extensions/mediatorMap/dsl/IMediatorMapper.ts
