+++ src/robotlegs/bender/extensions/mediatorMap/impl/ViewMediatorFactory.ts
