+++ src/robotlegs/bender/extensions/mediatorMap/impl/ViewMediatorManager.ts
