+++ src/robotlegs/bender/extensions/viewManager/SceneManagerExtension.ts
