+++ src/robotlegs/bender/extensions/viewManager/SceneManagerObserverExtension.ts
