+++ src/robotlegs/bender/extensions/viewManager/api/ISceneHandler.ts
