+++ src/robotlegs/bender/extensions/viewManager/api/ISceneManager.ts
