+++ src/robotlegs/bender/extensions/viewManager/api/IViewHandler.ts
