+++ src/robotlegs/bender/extensions/viewManager/impl/ConfigureSceneEvent.ts
