+++ src/robotlegs/bender/extensions/viewManager/impl/SceneManagerBinding.ts
