+++ src/robotlegs/bender/extensions/viewManager/impl/SceneManagerBindingEvent.ts
