+++ src/robotlegs/bender/extensions/viewManager/impl/SceneManagerObserver.ts
