+++ src/robotlegs/bender/extensions/viewManager/impl/SceneRegistry.ts
