+++ src/robotlegs/bender/extensions/viewManager/impl/SceneRegistryEvent.ts
